+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - added `FrmbFile` class
 - added `read_menu_hierarchy_as_file` function
+- added `enabled` attribute to `FrmbFormat`
 - test for `FrmbFormat` hashing 
 - changelog visible in documentation